PHP Session Timeout: ping server regularly
It is based on sending a save message to the server even if
there are no changes to be sent. This kills two birds with
one stone: it keeps the PHP session active and it does an
auto-save of changes which otherwise would have to be saved
manually (as described in another thread, changes made to
Tiddler tags are not save automatically).

If the ping is empty, then no notifications are generated
because that would be distracting. If changes are saved, this
is announced normally. I removed the "no changes saved" message
in the code which handles the reply: as far as I can tell,
this code was never triggered without the patch, but now it is
reached for empty pings and would create an unwanted pop-up.

The delay between pings is currently hard-coded to 1 minute.
I suppose this can (and should) be made configurable in a cookie.