ignore errors for a failed ping
When no changes had to be transmitted and a save message was sent
just to ping the server, then previously this was treating like
an error during saving:
- error message (both alert and text box)
- the full document was saved the next time

This overwrites all other changes made from a different computer
and such (usually transient) errors are of no interest to the user.
Now errors for a ping are silently ignored; errors for messages
which actually had to succeed are still displayed and handled
as before.